requirements.txt
# Rubrik CDM Pre-Upgrade Assessment — Dependencies
# Install: pip install -r requirements.txt

# HTTP requests
requests>=2.28.0,<3.0.0

# Environment variable management
python-dotenv>=1.0.0,<2.0.0

# SSL warning suppression (bundled with requests but pinned for clarity)
urllib3>=1.26.0,<3.0.0